+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  collectCoverage: true,
+  coverageDirectory: 'coverage',
+  coverageReporters: ['lcov', 'text'],
+  watchman: false
+}