+++ rollup.config.js
@@ -1,16 +1,30 @@
 
-var buble = require('rollup-plugin-buble')
-var external = Object
-    .keys(require('./package.json').dependencies)
-    .concat('fs', 'path')
+import buble from 'rollup-plugin-buble'
 
-module.exports = {
-  entry: 'src/index.js',
-  plugins: [buble()],
-  external: external,
+const external = Object
+  .keys(require('./package.json').dependencies)
+  .concat('fs', 'path')
+
+const banner = [
+  '/**',
+  ' * rollup-plugin-cleanup v' + require('./package.json').version,
+  ' * @author aMarCruz',
+  ' * @license MIT',
+  ' */'
+].join('\n')
+
+export default {
+  input: 'src/index.js',
+  plugins: [
+    buble({
+      target: { node: 4 }
+    })
+  ],
+  banner,
+  external,
   interop: false,
-  targets: [
-    { dest: 'dist/rollup-plugin-cleanup.js', format: 'cjs' },
-    { dest: 'dist/rollup-plugin-cleanup.es.js', format: 'es' }
+  output: [
+    { file: 'dist/rollup-plugin-cleanup.js', format: 'cjs' },
+    { file: 'dist/rollup-plugin-cleanup.es.js', format: 'es' }
   ]
 }