+++ src/cleanup.js
@@ -5,26 +5,29 @@ import blankComments from './blank-comments'
 import removeLines from './remove-lines'
 
 export default function cleanup(source, file, options) {
-  let changes
-  let code
 
-  if (options.comments === true) {
-    code = source
-  } else {
-    code = blankComments(source, file, options)
-    changes = code !== source
-  }
+  return new Promise(resolve => {
+    let changes
+    let code
 
-  const magicStr = new MagicString(code)
-
-  changes = removeLines(magicStr, code, file, options) || changes
-
-  if (changes) {
-    return {
-      code: magicStr.toString(),
-      map: options.sourceMap ? magicStr.generateMap({ hires: true }) : null
+    if (options.comments === true) {
+      code = source
+    } else {
+      code = blankComments(source, file, options)
+      changes = code !== source
     }
-  }
 
-  return null     // tell to Rollup that discard this result
+    const magicStr = new MagicString(code)
+
+    changes = removeLines(magicStr, code, file, options) || changes
+
+    resolve(changes
+      ? {
+        code: magicStr.toString(),
+        map: options.sourceMap ? magicStr.generateMap({ hires: true }) : null
+      }
+      : null)   // tell to Rollup that discard this result
+
+  })
+
 }